added team choice, master peer starting game, ugly though

+++ Test/start.cpp
@@ -41,8 +41,8 @@ void Start::mainMenu(){
 
 	NetworkManager::sInstance->SetState(NetworkManager::sInstance->NMS_MainMenu);
 	
-	queue.AddObject(menuButtons.Spawn(OPTIONS_BUTTON, (float)w - 125, (float)h - 250));
-	queue.AddObject(menuButtons.Spawn(JOIN_BUTTON, (float)w - 125, (float)h - 175));
+	queue.AddObject(menuButtons.Spawn(OPTIONS_BUTTON, w - 125, h - 250));
+	queue.AddObject(menuButtons.Spawn(JOIN_BUTTON, w - 125, h - 175));
 
 	while (NetworkManager::sInstance->GetState() == NetworkManager::sInstance->NMS_MainMenu){
